+++ PupuseriaSalvadorena/Repositorios/Interfaces/IDeclaracionTaxRep.cs
@@ -4,7 +4,7 @@ namespace PupuseriaSalvadorena.Repositorios.Interfaces
 {
     public interface IDeclaracionTaxRep
     {
-        Task CrearDeclaracionTax(int CedulaJuridica, DateTime FechaInicio, DateTime FechaFinal, decimal MontoTotalIngresos, decimal MontoTotalEgresos, decimal MontoTotalImpuestos, string Observaciones);
+        Task CrearDeclaracionTax(long CedulaJuridica, DateTime FechaInicio, DateTime FechaFinal, decimal MontoTotalIngresos, decimal MontoTotalEgresos, decimal MontoTotalImpuestos, string Observaciones);
         Task ActualizarDeclaracionTax(string IdDeclaracionImpuesto, DateTime FechaInicio, DateTime FechaFinal, decimal MontoTotalIngresos, decimal MontoTotalEgresos, decimal MontoTotalImpuestos, string Observaciones);
         Task EliminarDeclaracionTax(string IdDeclaracionImpuesto);
         Task<List<DeclaracionImpuesto>> MostrarDeclaracionesImpuestos();
