+++ PupuseriaSalvadorena/Repositorios/Interfaces/INegociosRep.cs
@@ -4,10 +4,10 @@ namespace PupuseriaSalvadorena.Repositorios.Interfaces
 {
     public interface INegociosRep
     {
-        Task CrearNegocio(int CedulaJuridica, string NombreEmpresa, int IdDireccion, int IdTelefono);
-        Task ActualizarNegocio(int CedulaJuridica, string NombreEmpresa);
-        Task EliminarNegocio(int CedulaJuridica);
+        Task CrearNegocio(long CedulaJuridica, string NombreEmpresa, int IdDireccion, int IdTelefono);
+        Task ActualizarNegocio(long CedulaJuridica, string NombreEmpresa);
+        Task EliminarNegocio(long CedulaJuridica);
         Task<List<Negocio>> MostrarNegocio();
-        Task<Negocio> ConsultarNegocio(int CedulaJuridica);
+        Task<Negocio> ConsultarNegocio(long CedulaJuridica);
     }
 }
